move external-secrets related types into external-secrets.ee folder
