+++ packages/cli/src/external-secrets.ee/errors/external-secrets-provider-not-found.error.ts
