+++ packages/cli/src/external-secrets.ee/errors/unknown-auth-type.error.ts
